// web/src/gen/model/userDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IEnumerableOfGuid } from "./iEnumerableOfGuid";
import type { IEnumerableOfString } from "./iEnumerableOfString";
import type { UserStatus } from "./userStatus";

export interface UserDto {
  id: string;
  userName: string;
  email: string;
  firstName: string;
  lastName: string;
  displayName: string;
  status: UserStatus;
  phoneNumber: string;
  createdAt: string;
  roles: IEnumerableOfString;
  roleIds: IEnumerableOfGuid;
}